+++ script/DeployUniswapV4AcrossMigrator.s.sol
@@ -29,7 +29,7 @@ contract DeployUniswapV4AcrossMigrator is Script, ChainSettlerHelper {
         );
 
         // set the chain settler addresses
-        (uint32[] memory chainIds, address[] memory chainSettlers, bool[] memory values) =
+        (uint256[] memory chainIds, address[] memory chainSettlers, bool[] memory values) =
             ChainSettlerHelper.getChainSettlersArrays("DEPLOY_CHAIN_IDS");
         if (chainIds.length > 0) {
             migrator.setChainSettlers(chainIds, chainSettlers, values);
